1. Replace Gson with better performant Jackson 2. Renamed Session.removeEventEventListener to removeEventListener. Old method is left as deprecated for backward compatibility 3. Session.navigate is marked as deprecated as the following commands may fail because of browserContext switch 4. Launcher now implements closeable interface and kills the browser process 5. SessionFactory got optional closeBrowserOnFactory close attribute that closes the browser when sessionFactory is closed 6. ObjectSubtypeHint got new possible values according to tip-of-tree 7. Launcher default constructor now chooses random port. Using same port may lead to failures when relaunching browser. 8. Some stability issues fixed 9. Few misspellings fixed 10. Other minor fixes

+++ src/test/java/io/webfolder/cdp/sample/ExecuteJavascript.java
@@ -30,7 +30,7 @@ public class ExecuteJavascript {
                             Session session = factory.create()) {
 
             session.waitDocumentReady();
-            Double result = (Double) session.evaluate("var foo = function() { return 2 + 2; }; foo();");
+            Integer result = (Integer) session.evaluate("var foo = function() { return 2 + 2; }; foo();");
             System.out.println(result);
 
             session.evaluate("var bar = {}; bar.myFunc = function(s1, s2) { return s1 + ' ' + s2; }");
